Giving up on GHC 7.8 in regards of Generic instances
Seems that GHC 7.8 has some problems with deriving instances for data
families. The simplest solution is to just not to derive problematic
instances on GHC 7.8. There might be a combination that would work for
all GHC versions, or maybe two variants that can be switched, but it
would make existing code in to a real mess.