ConnectionPool data family moved in to its own module
ConnectionPool data family was moved in to its own module named
Data.ConnectionPool.Family and it is now part of stable API.

This is a bacwards compatible change, since it is imported and
re-exported by Data.ConnectionPool.Internal.ConnectionPoolFamily module.

Problems with Typeable instance turn out to be problems with PolyKinds,
DeriveDataTypeable and TypeFamilies not being compatible language
extensions. In other words GHC versions <7.10 aren't able to derive it
for poly-kinded data family.